create pharmacy service class and tranfer allergy warning feature to it.
